Moving `showDialog` functions to `utils/helpers`
